Utils: Actually really fixes MetaClassResolver for PySide6
Turns out previous solution was broken. As we were intercepting the __new__ call before it reached type(QObject), and replacing the class argument to swap it with another one, it turns out this new class became the metaclass of the generated class.

So, it seemed to work fine for the class we were generating. But any subclass of that one was being created using the "wrong" (ie. incomplete) metaclass. And so the original metaclasses behaviour (eg. ABCMeta, SingletongMeta) were not honored for subclasses...

New solution seems to fixes it all. And even seems to be less hacky (but explanation for it is much more obscur though!).