(WIP) Nodes and Explorers, phase 1
This is ~2 years of (discontinuous) work. Reached the following implementation level:

- Essentials of Node, Children, ChildrenFactory, EntrySupport(Default), ChildrenStorage (aka ChildrenArray), GenericNode (aka AbstractNode), SyncChildren.

These were implemented very much as a bare translation from Netbeans Java code, with just minor adaptations wrt. how Python does some low-level things differently than Java.

This bare Java-translation was chosen to keep most of the logic, and get familiar with how Nodes functions internally. These are not meant to be kept as is (and so they are kept in a "_like_netbeans" subfolder). In phase 2 it will undergo a "Pythonisation". As well as shaving of deprecated/old parts that remained in the Java Netbeans version (for backward compatiblity reasons I guess, which we do not need to follow).

- A beginning of implementation of Explorer views. For these ones, instead of doing a bare Java-translation, it was more meaningful to start off from what Qt offers, ie. their Model/View framework.

A NodeModel was implementated, linking the Qt world with the Node components. It supports the hiearchy tree aspect of Nodes, in read-only mode (from Qt side), but modifiable (children added, removed, moved and node property change) from Node side (ie. using NodeListener).
It also delegate basic data access to Nodes properties display_name and icon.

Few Node-centric API were also implemented, notably for doing Node selection, as well as collapsing/expanding a Node children.
There are also more generic Node-to/from-Qt index helper methods.

- Finally, a bare Java-translation of Netbeans Mutex class. I understand it is a bit historical and predate nicer concurrency thing in JVM. But many things seems to relly on it that it seemed preferable to just reimplement it as-is for now (especially seeing the relative complexity of it!...).