FIX: EntrySupportDefault was missing up order of nodes on addition
The original algo was doing an it.remove(), where it was an iterator on a list. This got wrongly translated into python as in java it removes the last element it iterated over, whereas in python there is not such iterators, and its list.remove() remove the first occurrence in the list (removing the last wouldn't had been strictly the same neither, as it's not necessarily the same element than the last one we just got from the iterator, also here I think it would have somehow worked as in such cases it would have done it like that for all the matching elements).

Anyway, solution is to use enumerate, and pop the current index.

This was weirdly apparent during a ChildrenArray add test, where after adding a second time the same element, the internal order was getting messed up, but it was not getting immediately apparent. It only started to give out strange bug after adding 2 more different nodes. Then the order was complete non-sense...