Fixes bugs #3 and #4
The trick was to concentrate notifications only in Node._assign_to() and Node.deassign_from().

But, that was not that simple as filtering out the spurious notifications while not filtering out the actual legit notification (ie. when a parent node was really assigned to or deassigned from) was very tricky.

This is because of a few things:
- These methods are called very eagerly/extraneously by various code paths (notably in entry support) to ensure the state is consistent. Which lead to being called often when it's not really needed.
- But it is also difficult to determine when there is really a change of parent node from inside these methods because these methods are concerned by the assignment of a parent children, which may or may not have a parent node at the time of calling.
- And the way they are called (notably from Children._attach_to() and Children._detach_from()) makes it unreliable to ask the children what's its current parent node as these Children methods have already changed their internal state by the time they call those Node methods...

So, in Node._assign_to() we either filter on if we have a new parent children and the children has a parent node. Or, we now have an extra optional boolean parameter that says if this is called in case the children actually received a new parent node (so, only to be set by Children_attach_to()).

And in Node._deassign_from() it is slightly easier as we just need to care to have a reliable information about the old parent node (in order for the old == new filter in _fire_own_property_change to pass or not). When it is called by entry support, just getting the parent node of the old children is fine. But for when it is called by Children._detach_from() itself, now we also have an extra optional parameter for Children to pass us its old parent node directly.