Fixes bug #5
Decided that prev_snapshot should always be the previous snapshot and not some weird shortcut to one or the other snapshot.

And that when it comes to indices, ones should just be aware of if the event is for add or remove to pick the right snapshot to compare to.